+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Extracted the symbol search to a dedicated component
 - Upgraded `prisma` from version `4.14.1` to `4.15.0`
 
 ## 1.280.1 - 2023-06-10